requirements.txt
fastapi
uvicorn
python-dotenv
supabase
psycopg2-binary
python-jose